Question Bank

+++ database/migrations/2023_03_27_215137_create_payments_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
                 ->constrained('subscribes')
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
-            $table->double('ammount');
+            $table->decimal('ammount');
             $table->date('date');
             $table->timestamps();
         });
